requirements.txt
fastapi
uvicorn
asyncpg
mysql-connector-python
cachetools
crewai
langgraph
langchain-mistralai
python-dotenv
langchain
pydantic
pydantic-settings
cachetools
